+++ src/test/kotlin/com/zepben/evolve/streaming/get/testservices/TestCustomerConsumerService.kt
@@ -9,17 +9,33 @@
 package com.zepben.evolve.streaming.get.testservices
 
 import com.zepben.protobuf.cc.*
+import com.zepben.protobuf.metadata.GetMetadataRequest
+import com.zepben.protobuf.metadata.GetMetadataResponse
+import io.grpc.Status
 import io.grpc.stub.StreamObserver
 
 class TestCustomerConsumerService : CustomerConsumerGrpc.CustomerConsumerImplBase() {
 
     lateinit var onGetIdentifiedObjects: (request: GetIdentifiedObjectsRequest, response: StreamObserver<GetIdentifiedObjectsResponse>) -> Unit
-
     lateinit var onGetCustomersForContainer: (request: GetCustomersForContainerRequest, response: StreamObserver<GetCustomersForContainerResponse>) -> Unit
+    lateinit var onGetMetadataRequest: (request: GetMetadataRequest, response: StreamObserver<GetMetadataResponse>) -> Unit
+
     override fun getIdentifiedObjects(response: StreamObserver<GetIdentifiedObjectsResponse>): StreamObserver<GetIdentifiedObjectsRequest> =
         TestStreamObserver(response, onGetIdentifiedObjects)
 
     override fun getCustomersForContainer(responseObserver: StreamObserver<GetCustomersForContainerResponse>?): StreamObserver<GetCustomersForContainerRequest> =
         TestStreamObserver(responseObserver!!, onGetCustomersForContainer)
 
+    override fun getMetadata(request: GetMetadataRequest, responseObserver: StreamObserver<GetMetadataResponse>) {
+        runGrpc(request, responseObserver, onGetMetadataRequest)
+    }
+
+    private fun <T, U : StreamObserver<*>> runGrpc(request: T, response: U, handler: (T, U) -> Unit) {
+        try {
+            handler(request, response)
+            response.onCompleted()
+        } catch (t: Throwable) {
+            response.onError(Status.ABORTED.withDescription(t.message).asRuntimeException())
+        }
+    }
 }